requirements.txt
Django==4.0.6
djangorestframework==3.13.1
onesignal-sdk==2.0.0